Fix data type of osync_group_get_conflict_resolution' paramter "winner" to long long int, which represents a mmeber id (which is long long int)
refs #974



git-svn-id: http://svn.opensync.org/trunk@5456 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e

+++ opensync/group/opensync_group_internals.h
@@ -52,7 +52,7 @@ struct OSyncGroup {
 	/** The defined resolution for this group */
 	OSyncConflictResolution conflict_resolution;
 	/** The winning side if the select resolution is choosen */
-	int conflict_winner;
+	long long int conflict_winner;
 	/** The configured merger status of this group */
 	osync_bool merger_enabled;
 	/** The configured converter status of this group */
